requirements.txt
google-cloud-bigquery
pandas
db-dtypes
google-auth-oauthlib
python-dotenv
tqdm
